moved the plugin interface into the tmuxp package instead of its own separate package
