--- tests/fixtures/pluginsystem/plugins/tmuxp_plugin_interface/setup.py
@@ -1,10 +1,0 @@
-from setuptools import setup
-
-setup(
-    name='tmuxp_plugin_interface',
-    version='0.0.1',
-    author='Joseph Flinn',
-    author_email='joseph.s.flinn@gmail.com',
-    packages=setuptools.find_packages(),
-    description='A tmuxp plugin class interface',
-)
