+++ tests/fixtures/pluginsystem/plugins/tmuxp_test_plugin_bs/tmuxp_test_plugin_bs/plugin.py
@@ -1,4 +1,4 @@
-from tmuxp_plugin_interface.plugin import TmuxpPluginInterface
+from tmuxp.plugin import TmuxpPluginInterface
 
 class PluginBeforeScript(TmuxpPluginInterface):
     def __init__(self):
