+++ tests/fixtures/pluginsystem/plugins/tmuxp_test_plugin_bwb/setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name='tmuxp_test_plugin_bwb',
-    version='0.0.1',
+    version='0.0.2',
     author='Joseph Flinn',
     author_email='joseph.s.flinn@gmail.com',
     packages=setuptools.find_packages(),
